driver/ubootdriver: port idle console handling from BareboxDriver
Activating the UBootDriver for a U-Boot that is already interrupted does
not work at the moment, because nothing is received on the console since
nothing is sent in the first place.

To change that, harmonize its _await_prompt() method with the one found
in BareboxDriver. BareboxDriver already implements handling the idle
console functionality.

This allows a strategy to force() a U-Boot state.

Signed-off-by: Bastian Krause <bst@pengutronix.de>